requirements.txt
streamlit>=1.31.0
yfinance>=0.2.36
pandas>=2.2.2
numpy>=2.1.0
plotly>=5.18.0
requests>=2.31.0
transformers>=4.37.2
torch>=2.2.0
scikit-learn>=1.4.0
ta>=0.11.0
nltk>=3.8.1
python-dotenv>=1.0.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.3
feedparser>=6.0.11
textblob>=0.18.0
pytz>=2024.1